requirements.txt
PySide6>=6.5.0
watchdog>=3.0.0
python-telegram-bot[asyncio]>=20.0
loguru>=0.7.0
tomli>=2.0.1
pytest>=7.0.0
